Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/module-scriptnonce-allowed.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Security-Policy" content="script-src 'nonce-noncynonce' 'nonce-noncy+/=nonce'">
        <script nonce="noncynonce">
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
        }

        function done(msg) {
            document.querySelector("pre").innerHTML = msg;
            if (window.testRunner)
                testRunner.notifyDone();
        }
        </script>
        <script type="module" nonce="noncynonce">
            alert('PASS (1/3)');
        </script>
        <script type="module" nonce="noncynonce">
            alert('PASS (2/3)');
        </script>
        <script type="module" nonce="noncy+/=nonce">
            alert('PASS (3/3)');
        </script>
        <script type="module" nonce="noncynonce">
            done("PASS");
        </script>
    </head>
    <body>
        <p>
            This tests the effect of a valid module script-nonce value. It passes if
            no console warning is visible and the alerts are executed.
        </p>
        <pre></pre>
    </body>
</html>
