Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/module-scriptnonce-allowed-expected.txt

ALERT: PASS (1/3)
ALERT: PASS (2/3)
ALERT: PASS (3/3)
This tests the effect of a valid module script-nonce value. It passes if no console warning is visible and the alerts are executed.

PASS
